Fully use esp_ableton_link as a remote component
Local bindings file includes header from remote component, this way we
don't have to update for any changes in Link

Also, apparently `version` is needed on remote components, for
esp-idf-sys to pick it up correctly. Without it, it silently fails to
parse the component, and the bindings are not generated.